more code structure moving
